refactor: :recycle: move the preparation providers out of the widget file
`detailed_preparation_widget.dart` declared three providers alongside the
widget, so the screen's state lived in a file named after one of the
widgets that reads it — and the screen had to import a widget to reach
`preparationProgressProvider`. Every other provider in the app sits in
`presentation/providers/`.

They now live in `preparation_providers.dart`, with a note on why these
three are autoDispose while the rest of the app is keepAlive: they
describe an open screen and are meant to reset when it closes.

Verified on the emulator, two players through to Preparation: the global
progress bar advances as steps are ticked, per-phase counters follow
(2/4), and collapsing and expanding a phase still works.
