+++ lib/features/game_setup/presentation/screens/game_setup_preparation_screen.dart
@@ -2,6 +2,7 @@ import 'package:companion_for_cacao/core/theme/app_colors.dart';
 import 'package:companion_for_cacao/features/game_setup/domain/entities/game_setup_state_entity.dart';
 import 'package:companion_for_cacao/features/game_setup/presentation/providers/game_setup_notifier.dart';
 import 'package:companion_for_cacao/features/game_setup/presentation/providers/preparation_view_mode_provider.dart';
+import 'package:companion_for_cacao/features/game_setup/presentation/providers/preparation_providers.dart';
 import 'package:companion_for_cacao/features/game_setup/presentation/widgets/detailed_preparation_widget.dart';
 import 'package:companion_for_cacao/features/game_setup/presentation/widgets/guided_preparation_widget.dart';
 import 'package:companion_for_cacao/l10n/generated/app_localizations.dart';
